+++ nobinobi_core/views.py
@@ -11,7 +11,7 @@
 #
 #      You should have received a copy of the GNU Affero General Public License
 #      along with this program.  If not, see <https://www.gnu.org/licenses/>.
-
+#
 # -*- coding: utf-8 -*-
 
 import arrow
@@ -20,10 +20,7 @@ from django.contrib import admin, messages
 from nobinobi_core.forms import AddOfficialHolidayForm
 from nobinobi_core.models import Holiday
 
-try:
-    from django.urls import reverse
-except:
-    from django.core.urlresolvers import reverse
+from django.urls import reverse
 from django.db import IntegrityError
 from django.http import HttpResponseRedirect
 from django.views.generic import FormView