+++ HISTORY.rst
@@ -3,7 +3,14 @@
 History
 -------
 
-0.1.0 (2018-10-17)
+0.1.1 (2020-10-06)
+++++++++++++++++++
+
+* Add company & Company closure
+* Restruct package for better comprehension
+* Add tests
+
+0.1.0 (2020-10-06)
 ++++++++++++++++++
 
 * First release on PyPI.